+++ src/scripts/items/in_wall_floor_item.js
@@ -1,7 +1,3 @@
-/// <reference path="../../lib/three.d.ts" />
-/// <reference path="../model/model.ts" />
-/// <reference path="in_wall_item.ts" />
-/// <reference path="metadata.ts" />
 import {InWallItem} from './in_wall_item.js';
 
 /** */